Create Level, refactor ModelManager and StationState.
I disagree that StationState should be the initial state of the Level. This is because Level needs to contain more information such as `levelDescriptor` and `expectedOutput`. Currently, `expectedOutput` is placed in StationState. However, it should not be there as in ModelManager, the undoStack: Stack<StationState>, and as such it doesn’t make sense for ModelManager to store repeated values of `expectedOutput` in the stack i.e it is repetitive.

Also, ModelManager should not have the `stationName`, it should be stored in `Level` instead.