refactor `isExecutionAllowed` from `LogicManager` to `RunState`.
`isExecutionAllowed` is true when there is no animation going on. However,
this logic should rather be kept in the `RunState` instead as it refers to
the state of the game (whether it is animating or not).

In the current approach, it’s weird that `LogicManager` should know whether
the execution of a command causes the person to move or not. As such,
`GameScene` should rather be the class notifying when animation begins
or ends.